+++ src/shoffing/expression/OperatorTan.java
@@ -9,7 +9,7 @@ public class OperatorTan extends Operator
 		if(subTree.getLeft() != null)
 		{
 			double value = subTree.getLeft().getRoot().eval( vars, subTree.getLeft() );
-			return (double) Math.tan(value);
+			return  Math.tan(value);
 		} else
 			return 0;
 	}
